+++ eslint.config.js
@@ -1,3 +1,4 @@
+// @ts-check
 import eslintConfig from '@antfu/eslint-config'
 
 export default eslintConfig({
