+++ src/core/server/filePathMap.ts
@@ -1,0 +1,6 @@
+import { defineEventHandler } from 'h3'
+
+export default defineEventHandler(async () => {
+  const filePathMap = globalThis.TurboConsoleFilePathMap || new Map()
+  return Object.fromEntries(filePathMap)
+})
