+++ src/core/server/health.ts
@@ -1,0 +1,7 @@
+import { defineEventHandler } from 'h3'
+
+export default defineEventHandler(async () => {
+  return {
+    status: 'success',
+  }
+})
