+++ src/core/server/launchEditor.ts
@@ -1,0 +1,40 @@
+import { cwd } from 'node:process'
+import { defineEventHandler, getQuery } from 'h3'
+// @ts-expect-error no types
+import launch from 'launch-editor'
+import { resolve } from 'pathe'
+
+export default defineEventHandler(async (event) => {
+  try {
+    const { position } = getQuery(event) as { position: string }
+    const filePathMap = globalThis.TurboConsoleFilePathMap
+
+    if (!filePathMap)
+      throw new Error('filePathMap is undefined')
+
+    const parsed = position.split(',')
+    const filePathMapString = parsed[0]
+    const line = parsed[1] || 1
+    const column = parsed[2] || 1
+
+    let filePath = ''
+
+    for (const [key, value] of filePathMap) {
+      if (value === filePathMapString) {
+        filePath = key
+        break
+      }
+    }
+
+    launch(resolve(cwd(), `${filePath}:${line}:${column}`))
+    return {
+      status: 'success',
+    }
+  }
+  catch (error) {
+    return {
+      status: 'error',
+      message: String(error),
+    }
+  }
+})
