+++ src/globals.d.ts
@@ -4,4 +4,5 @@ export interface global {}
 
 declare global {
   var TurboConsoleFilePathMap: Map<string, string>
+  var UNPLUGIN_TURBO_CONSOLE_LAUNCH_SERVER: boolean
 }
