+++ test/filePathMap.test.ts
